requirements.txt
# Please use pip to install tensorflow-related tool.
tensorflow-gpu==1.15